Move the package into a single file
I was having issues with building on linux.  If my issue was actually
that I am using pyinstaller wrong, then this commit shouldn't be
accepted.

This observes that the package was previous three files:
- /b3denv.py, which just calls main() from b3denv.__main__
- /b3denv/__main__.py, which just calls main() from b3denv.__init__
- /b3denv/__init__.py, which defines all the logic (and incidentally
  also calls main())

That there's only one file that does more than call main() suggests that
this should actually be a single-source project.  This commit implements
that.
