--- b3denv/__main__.py
@@ -1,4 +1,0 @@
-from b3denv import main
-
-if __name__ == "__main__":
-    main()
